prints elements in a list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -4,11 +4,11 @@
  * print_list - prints elements of a list
  * @h : pointer to the head of the list
  *
- * Return: the size of elements in the node
+ * Return: the number of nodes
  */
 size_t print_list(const list_t *h)
 {
-	int s = 0;
+	size_t  s = 0;
 
 
 	while (h != NULL)
@@ -19,7 +19,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%d] %s\n" h->len, h->str);
+			printf("[%u] %s\n" h->len, h->str);
 		}
 		h = h->next;
 		s++;
